Added the delete conversation function
