+++ client/src/components/Conversation.jsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import Auth from '../utils/auth';
 import { useMutation } from '@apollo/client';
-import { ADD_COMMENT, DELETE_CONVERSATION } from '../utils/mutations'; 
+import { ADD_COMMENT } from '../utils/mutations'; 
 import { useQuery } from '@apollo/client';
 import { GET_CONVERSATION_BY_ID } from '../utils/queries';
 import './Conversation.css';
@@ -17,8 +17,6 @@ const Conversation = ({ onClose }) => {
   });
 
   const [addComment, { commentError }] = useMutation(ADD_COMMENT);
-
-  const [deleteConversation] = useMutation(DELETE_CONVERSATION);
 
   const handleChange = (event) => {
     const { value } = event.target;
@@ -67,21 +65,6 @@ const Conversation = ({ onClose }) => {
     }
   };
   
-  const handleEndConvo = async (event) => {
-    event.preventDefault();
-    console.log("Im reaching here - end convo")
-    try {
-      await deleteConversation({
-        variables: {
-          conversationId: conversationId,
-          username: Auth.getProfile().data.username
-        },
-      });
-    } catch (err) {
-      console.error(err);
-    }
-  };
-
   return (
     <div className="conversation-container">
       <div className="conversation-title">{fetchedConversation.conversationTitle}</div>
@@ -116,18 +99,6 @@ const Conversation = ({ onClose }) => {
         >
           Add Comment
         </button>
-        {fetchedConversation.isPrivate ? (
-            <button 
-            type="button" 
-            className="btn btn-primary end-convo-btn"
-            onClick={handleEndConvo}
-          >
-            End Conversation
-  
-          </button>
-        ) : (
-          <></>
-        )}
       </div>
     </div>
   );
