+++ client/src/components/UserConversation.jsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react';
 import Auth from '../utils/auth';
 import { useMutation } from '@apollo/client';
-import { ADD_COMMENT } from '../utils/mutations'; 
+import { ADD_COMMENT, DELETE_CONVERSATION } from '../utils/mutations'; 
 import { useQuery } from '@apollo/client';
 import { GET_CONVERSATION_BY_ID } from '../utils/queries';
 import { GET_USER_BY_ID } from '../utils/queries';
@@ -35,6 +35,8 @@ const UserConversation = ({ onClose }) => {
 
   const [addComment, { error: commentError }] = useMutation(ADD_COMMENT);
 
+  const [deleteConversation] = useMutation(DELETE_CONVERSATION);
+
   useEffect(() => {
     if (conversationId) {
       refetch();
@@ -62,6 +64,21 @@ const UserConversation = ({ onClose }) => {
       refetch();
     } catch (error) {
       console.error("Error adding comment:", error);
+    }
+  };
+
+  const handleEndConvo = async (event) => {
+    event.preventDefault();
+    console.log("Im reaching here - end convo")
+    try {
+      await deleteConversation({
+        variables: {
+          conversationId: conversationId,
+          username: Auth.getProfile().data.username
+        },
+      });
+    } catch (err) {
+      console.error(err);
     }
   };
 
@@ -108,6 +125,18 @@ const UserConversation = ({ onClose }) => {
         >
           Add Comment
         </button>
+        {fetchedConversation.isPrivate ? (
+            <button 
+            type="button" 
+            className="btn btn-primary end-convo-btn"
+            onClick={handleEndConvo}
+          >
+            End Conversation
+  
+          </button>
+        ) : (
+          <></>
+        )}
       </div>
     </div>
   );
